+++ ItemsDto/Barriers.h
@@ -13,10 +13,10 @@ namespace ItemsDto {
         [[nodiscard]] const vector<Cube *> &getItems() const override;
         void createWall(int x, int y);
         void reset();
-        [[nodiscard]] float getMaxX() const override;
-        [[nodiscard]] float getMaxY() const override;
-        [[nodiscard]] float getMinX() const override;
-        [[nodiscard]] float getMinY() const override;
+        [[nodiscard]] int getMaxX() const override;
+        [[nodiscard]] int getMaxY() const override;
+        [[nodiscard]] int getMinX() const override;
+        [[nodiscard]] int getMinY() const override;
 
     protected:
         static Cube* wallFactory(const glm::vec3 &position);
