+++ ItemsDto/ObjWall.h
@@ -11,10 +11,10 @@ namespace ItemsDto {
         ~ObjWall();
         void init() override;
         [[nodiscard]] const vector<ObjItem *> &getItems() const override;
-        [[nodiscard]] float getMaxX() const override;
-        [[nodiscard]] float getMaxY() const override;
-        [[nodiscard]] float getMinX() const override;
-        [[nodiscard]] float getMinY() const override;
+        [[nodiscard]] int getMaxX() const override;
+        [[nodiscard]] int getMaxY() const override;
+        [[nodiscard]] int getMinX() const override;
+        [[nodiscard]] int getMinY() const override;
     protected:
         void release() override;
         vector<ObjItem*> walls;
